Code optimization and CC flags

+++ srcs/renderer/render.c
@@ -6,7 +6,7 @@
 /*   By: plouvel <plouvel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/07 18:05:35 by plouvel           #+#    #+#             */
-/*   Updated: 2022/01/29 14:42:51 by plouvel          ###   ########.fr       */
+/*   Updated: 2022/01/29 16:04:24 by plouvel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,17 +78,17 @@ void	render(t_mlx *fdf)
 	t_vec2d	p1;
 
 	vec3d.y = 0;
-	while (vec3d.y < fdf->data.nbr_lines)
+	while ((size_t) vec3d.y < fdf->data.nbr_lines)
 	{
 		vec3d.x = 0;
-		while (vec3d.x < fdf->data.elems_line)
+		while ((size_t) vec3d.x < fdf->data.elems_line)
 		{
 			vec3d.z = fdf->data.vertices[vec3d.y][vec3d.x];
 			p1 = fdf->data.curr_proj->transform(fdf->data.curr_proj->tile_width,
 					fdf->data.curr_proj->org, vec3d, &fdf->data);
 			if ((p1.x > 200 && p1.x <= WIDTH) && (p1.y >= 0 && p1.y <= HEIGHT))
 			{
-				if (vec3d.x < fdf->data.elems_line - 1)
+				if ((size_t) vec3d.x < fdf->data.elems_line - 1)
 					compute_n_draw(fdf, p1, vec3d.x + 1, vec3d.y);
 				if (vec3d.y > 0)
 					compute_n_draw(fdf, p1, vec3d.x, vec3d.y - 1);
